Warning fixes

+++ src/common/Text/GXmlTreeUi.cpp
@@ -217,7 +217,7 @@ int GetDataType(char *str)
 		bool Float = false;
 
 		char16 w;
-		int Len = strlen(str);
+		int Len = (int)strlen(str);
 		while ((w = LgiUtf8To32((uint8*&)str, Len)))
 		{
 			if (strchr("e \t\r\n", w))
